Remove objc classes
