+++ OptimoveSDK/Sources/Classes/PreferenceCenter/Models/PreferenceUpdateRequest.swift
@@ -1,0 +1,11 @@
+import Foundation
+
+public struct PreferenceUpdateRequest: Codable {
+    public let topicId: String
+    public let channelSubscription: [Channel]
+
+    public init(topicId: String, channelSubscription: [Channel]) {
+        self.topicId = topicId
+        self.channelSubscription = channelSubscription
+    }
+}
